llm-cli-v4/backend/pyproject.toml
[project]
name = "llm-cli-v3"
version = "3.0.0"
description = "LLM CLI V3 - Web + CLI 双模式"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.6",
    "openai>=1.0.0",
    "PyYAML>=6.0.1",
    "pytz>=2023.3",
]

[project.optional-dependencies]
dev = [
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
]

[build-system]
requires = ["setuptools>=68.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 100
target-version = ["py39"]

[tool.isort]
profile = "black"
line_length = 100

[tool.pyproject.toml]
